requirements.txt
llama-cpp-python>=0.2.77
streamlit>=1.35.0
huggingface_hub>=0.23.0
